requirements.txt
torch==2.3.0
torchvision==0.18.0
faiss-cpu==1.8.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
pillow==10.4.0
matplotlib==3.9.0
typer[all]==0.12.3
rich==13.7.1
